requirements.txt
streamlit
langchain
openai
python-dotenv
yfinance
pandas
numpy
plotly
